requirements.txt
fastapi == 0.104.0
python-dotenv == 1.0.0
uvicorn == 0.23.2
pydantic == 2.4.2
requests == 2.31.0
motor == 3.3.1
schedule == 1.2.1
aioschedule == 0.5.2